# book-store/src/main/resources/openapi/book-model.yaml
openapi: 3.0.2
info:
  title: Book Store Api First model
  version: 1.1.0
  description: This API refers UI from 'Role Concept User Managament.pptx'

#================================================================================
#
# PATHS
#
#================================================================================
paths:
  /books:
    description: >
      TODO:
      - autocomplete
    get:
      tags:
        - books
      operationId: getAllBooks
      responses:
        '200':
          description: Returns all books
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Book'
        '403':
          $ref: '#/components/responses/403errorResponse'
    post:
      tags:
        - books
      operationId: addBook
      requestBody:
        description: Book object to add
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Book'
      responses:
        '201':
          description: Returns created book
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Book'
        '400':
          $ref: '#/components/responses/4xxerrorResponse'
    put:
      tags:
        - books
      operationId: updateBook
      requestBody:
        description: Update Book
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Book'
      responses:
        '200':
          description: Returns updated book
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Book'
        '400':
          $ref: '#/components/responses/4xxerrorResponse'
  /books/{uuid}:
    description: >
      TODO:
      - autocomplete
    parameters:
      - $ref: '#/components/parameters/uniqueUuid'
    get:
      tags:
        - books
      operationId: getBookByUuid
      responses:
        '200':
          description: Returns book by UUID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Book'
        '403':
          $ref: '#/components/responses/403errorResponse'
  /authors:
    description: >
      TODO:
      - autocomplete
    get:
      tags:
        - authors
      operationId: getAllAuthors
      responses:
        '200':
          description: Returns all authors
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Author'
        '403':
          $ref: '#/components/responses/403errorResponse'
    post:
      tags:
        - authors
      operationId: createAuthor
      requestBody:
        description: Author object to creat
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Author'
      responses:
        '200':
          description: Returns created author
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Author'
        '403':
          $ref: '#/components/responses/403errorResponse'
    put:
      tags:
        - authors
      operationId: updateAuthor
      requestBody:
        description: Update Author
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Author'
      responses:
        '200':
          description: Returns updated Author
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Author'
        '400':
          $ref: '#/components/responses/4xxerrorResponse'
  /authors/{uuid}:
    description: >
      TODO:
      - autocomplete
    parameters:
      - $ref: '#/components/parameters/uniqueUuid'
    get:
      tags:
        - authors
      operationId: getAuthorByUuid
      responses:
        '200':
          description: Returns Author by UUID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Author'
        '403':
          $ref: '#/components/responses/403errorResponse'
...
components:
  responses:
    4xxerrorResponse:
      description: Response used to provide error details
      content:
        application/json:
          example: '{ "error": "validation.name.unique" }'
          schema:
            type: object
            properties:
              error:
                type: string
    403errorResponse:
      description: Client don't have permissions to perform operation
  parameters:
...
    uniqueUuid:
      name: uuid
      in: path
      required: true
      schema:
        $ref: '#/components/schemas/Uuid'
...
  schemas:
    Uuid:
      type: string
      format: uuid
...
    BookDetails:
      type: object
      properties:
        title:
          type: string
        isbn:
          type: string
        description:
          type: string
        uuid:
          $ref: '#/components/schemas/Uuid'
    Book:
      type: object
      properties:
        bookDetails:
          $ref: '#/components/schemas/BookDetails'
        authorDetails:
          $ref: '#/components/schemas/AuthorDetails'
    AuthorDetails:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        description:
          type: string
        uuid:
          $ref: '#/components/schemas/Uuid'
    Author:
      type: object
      properties:
        authorDetails:
          $ref: '#/components/schemas/AuthorDetails'
        books:
          type: array
          items:
            $ref: '#/components/schemas/BookDetails'
...